requirements.txt
fastapi==0.111.0
fastapi-cli==0.0.4
pydantic==2.7.4
pydantic_core==2.18.4
scikit-learn==1.2.2
numpy==1.24.4
pandas==1.5.3
tensorflow==2.16.1
yfinance==0.2.22
python-dotenv==1.0.1
uvicorn==0.30.1